# docs/swagger.yaml
definitions:
...
  model.HTTPMessage:
    description: HTTPMessage represents a message to be sent to the client trough
      HTTP REST API.
    properties:
      message:
        example: success
        format: string
        type: string
      method:
        example: GET
        format: string
        type: string
      path:
        example: /api/v1/users
        format: string
        type: string
      status_code:
        example: 200
        format: int32
        type: integer
      timestamp:
        example: "2021-07-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  model.LoginUserRequest:
    description: LoginUserRequest is the request struct for the LoginUser handler.
    properties:
      email:
        example: admin@qu3ry.me
        format: email
        type: string
      password:
        example: ThisIsApassw0rd.,
        format: string
        type: string
    required:
    - email
    - password
    type: object
  model.LoginUserResponse:
    description: LoginUserResponse is the response when a user logs in.
    properties:
      access_token:
        format: string
        type: string
      permissions:
        additionalProperties: {}
        format: object
        type: object
      refresh_token:
        format: string
        type: string
      token_type:
        example: Bearer
        format: string
        type: string
      user_id:
        example: 01979cde-6d91-773b-a03c-3779be3b55b3
        format: uuid
        type: string
    type: object
...
  model.ReVerifyUserRequest:
    description: ReVerifyUserRequest is the request struct for the ReVerifyUser handler.
    properties:
      email:
        example: user@mail.com
        format: email
        type: string
    required:
    - email
    type: object
  model.RefreshTokenRequest:
    description: RefreshTokenRequest is the request struct for the RefreshToken handler.
    properties:
      refresh_token:
        format: string
        type: string
    type: object
  model.RefreshTokenResponse:
    description: RefreshTokenResponse is the response when a user refreshes their
      token.
    properties:
      access_token:
        format: string
        type: string
      token_type:
        example: Bearer
        format: string
        type: string
    type: object
  model.RegisterUserRequest:
    description: RegisterUserRequest is the request struct for the RegisterUser handler.
    properties:
      email:
        example: john.doe@email.com
        format: email
        type: string
      first_name:
        example: John
        format: string
        type: string
      id:
        example: 01979cde-6d91-773f-a630-7eed93217f35
        format: uuid
        type: string
      last_name:
        example: Doe
        format: string
        type: string
      password:
        example: ThisIsApassw0rd.,
        format: string
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
info:
  contact:
    email: info@qu3ry.me
    name: API Support
    url: https://qu3ry.me
  description: |-
    This is a service template for building RESTful APIs in Go.
    It uses a PostgreSQL database to store user information.
    The service provides:
    - CRUD operations for users.
    - Health and version endpoints.
    - Configuration using environment variables or command line arguments.
    - Debug mode to enable debug logging.
    - TLS enabled to secure the communication.
  title: Go REST API Service Template
  version: v1
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user credentials and return JWT access and refresh
        tokens
      operationId: 019791cc-06c7-7d8c-81e4-914dd89098e8
      parameters:
      - description: The information of the user to login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Logout user and invalidate session tokens
      operationId: 019791cc-06c7-7e4c-961c-b1bf5f40d633
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: Logout user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate new access token using valid refresh token
      operationId: 019791cc-06c7-7e50-a5b6-bd1c82d5c031
      parameters:
      - description: The refresh token to use
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - RefreshToken: []
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and send email verification
      operationId: 019791cc-06c7-7e38-ac82-6919df575ff7
      parameters:
      - description: The information of the user to register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      summary: Register user
      tags:
      - Auth
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Resend account verification email to user
      operationId: 019791cc-06c7-7e48-b046-99f072088c50
      parameters:
      - description: The email of the user to re-verify
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ReVerifyUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      summary: Resend verification
      tags:
      - Auth
  /auth/verify/{jwt}:
    get:
      description: Verify user account using JWT verification token
      operationId: 019791cc-06c7-7e40-bcc3-149cd13bee44
      parameters:
      - description: The JWT token to use
        format: jwt
        in: path
        name: jwt
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      summary: Verify user
      tags:
      - Auth
...
swagger: "2.0"